requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
importlib-metadata==5.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.9.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.33
protobuf==3.20.3
python-docx==0.8.11
python-dotenv==1.0.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
zipp==3.9.0
